Tweak variables. Trying to work out how to do player select.

+++ glstation.h
@@ -26,7 +26,7 @@ private:
   
   std::list<GLShip*>* objects, *targets;
   int ships_this_wave, max_ships_per_wave, extra_ships_per_wave, ships_left_to_deploy;
-  float time_until_next_ship, time_between_ships;
+  int time_until_next_ship, time_between_ships;
   bool deploying;
   int wave, difficulty;
 };
